requirements.txt
Django==5.2.5
reportlab==4.2.2
openpyxl==3.1.2
django-allauth==65.10.0
requests==2.32.4
PyJWT==2.10.1
cryptography==45.0.6